+++ packages/typespec-autorest/src/openapi.ts
@@ -10,6 +10,7 @@ import {
   hasUniqueItems,
 } from "@azure-tools/typespec-azure-core";
 import {
+  ArmFeatureOptions,
   getArmCommonTypeOpenAPIRef,
   getArmIdentifiers,
   getArmKeyIdentifiers,
@@ -150,7 +151,6 @@ import {
   resolveInfo,
 } from "@typespec/openapi";
 import { getVersionsForEnum } from "@typespec/versioning";
-import { ArmFeatureOptions } from "../../typespec-azure-resource-manager/generated-defs/Azure.ResourceManager.Legacy.js";
 import { AutorestOpenAPISchema } from "./autorest-openapi-schema.js";
 import { getExamples, getRef } from "./decorators.js";
 import { sortWithJsonSchema } from "./json-schema-sorter/sorter.js";