+++ packages/typespec-azure-portal-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog - @azure-tools/typespec-azure-portal-core
 
+## 0.65.0
+
+No changes, version bump only.
+
 ## 0.64.0
 
 ### Bump dependencies